LayoutTests/platform/chromium/http/tests/workers/worker-importScriptsOnError-expected.txt

CONSOLE MESSAGE: line 2: Uncaught undefined
Test importScripts with error.


AT_TARGET: 2,
BLUR: 8192,
BUBBLING_PHASE: 3,
CAPTURING_PHASE: 1,
CHANGE: 32768,
CLICK: 64,
DBLCLICK: 128,
DRAGDROP: 2048,
FOCUS: 4096,
KEYDOWN: 256,
KEYPRESS: 1024,
KEYUP: 512,
MOUSEDOWN: 1,
MOUSEDRAG: 32,
MOUSEMOVE: 16,
MOUSEOUT: 8,
MOUSEOVER: 4,
MOUSEUP: 2,
NONE: 0,
SELECT: 16384,
bubbles: false,
cancelBubble: false,
cancelable: true,
clipboardData: undefined,
currentTarget: [object Worker],
defaultPrevented: false,
eventPhase: 2,
filename: http://127.0.0.1:8000/workers/resources/worker-importScripts-error.js,
initEvent: function initEvent() { [native code] },
lineno: 2,
message: Uncaught undefined,
preventDefault: function preventDefault() { [native code] },
returnValue: true,
srcElement: [object Worker],
stopImmediatePropagation: function stopImmediatePropagation() { [native code] },
stopPropagation: function stopPropagation() { [native code] },
target: [object Worker],
type: error
